+++ packages/commercetools/api-client/src/api/createCart/defaultMutation.ts
@@ -4,7 +4,7 @@ import { CartFragment } from './../../fragments';
 export default gql`
   ${CartFragment}
 
-  mutation createCart($draft: MyCartDraft!, $locale: Locale!, $acceptLanguage: [Locale!], $storeKey: KeyReferenceInput) {
+  mutation createCart($draft: MyCartDraft!, $locale: Locale!, $acceptLanguage: [Locale!], $currency: Currency!, $storeKey: KeyReferenceInput) {
     cart: createMyCart(draft: $draft, storeKey: $storeKey) {
       ...DefaultCart
     }