+++ packages/core/core/src/types.ts
@@ -1,7 +1,21 @@
 /* istanbul ignore file */
-
 import { Ref } from '@vue/composition-api';
 import type { Request, Response } from 'express';
+
+/**
+ * Default name of the cookie storing active localization code
+ */
+export const VSF_LOCALE_COOKIE = 'vsf-locale';
+
+/**
+ * Default name of the cookie storing active currency code
+ */
+export const VSF_CURRENCY_COOKIE = 'vsf-currency';
+
+/**
+ * Default name of the cookie storing active country code
+ */
+export const VSF_COUNTRY_COOKIE = 'vsf-country';
 
 export type ComputedProperty<T> = Readonly<Ref<Readonly<T>>>;
 
@@ -60,7 +74,7 @@ export interface UseUser
   register: (params?: { user: UseUserRegisterParams }) => Promise<void>;
   login: (params?: { user: UseUserLoginParams }) => Promise<void>;
   logout: () => Promise<void>;
-  changePassword: (params: { current: string; new: string }) => Promise<void>;
+  changePassword: (params?: { current: string; new: string }) => Promise<void>;
   load: () => Promise<void>;
   isAuthenticated: Ref<boolean>;
   loading: ComputedProperty<boolean>;